refactor: Cars, CarFactory, NumberGenerator, CarTest, RacingCarTest
