+++ src/main/java/racingcar/util/NumberGenerator.java
@@ -1,8 +1,5 @@
 package racingcar.util;
 
 public interface NumberGenerator {
-    int MAX = 9;
-    int MIN = 0;
-
     int generateNumber();
 }
