+++ src/test/java/racingcar/TestNumberGenerator.java
@@ -3,8 +3,8 @@ package racingcar;
 import racingcar.util.NumberGenerator;
 
 public class TestNumberGenerator implements NumberGenerator {
-    int[] numbers;
-    int index = 0;
+    private final int[] numbers;
+    private int index = 0;
 
     public TestNumberGenerator(int[] numbers) {
         this.numbers = numbers;
